second commit

+++ Task_1/src/main/java/com/epam/seliazniova/entity/Point.java
@@ -4,25 +4,25 @@ package com.epam.seliazniova.entity;
  * Created by Iryna_Seliazniova on 6/12/2017.
  */
 public class Point {
-    private int x;
-    private int y;
-    private int z;
+    private double x;
+    private double y;
+    private double z;
 
-    public Point(int x, int y, int z) {
+    public Point(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public int getZ() {
+    public double getZ() {
         return z;
     }
 }
